Refactor Check and X buttons into SelectionButton
