+++ src/lib/components/buttons/selection-button-type.ts
@@ -1,0 +1,4 @@
+export enum SelectionButtonType {
+	Check,
+	X,
+}
